Allow to queue client requests if client is busy
This should fix https://github.com/Openwsman/openwsman/issues/10

If Openwsman client is run in a multithreaded fashion, client requests
(== any wsmc_action_* call) can be fired simultaneously.

The original implementation rejected (returned error value) if a new
request was issued while the current is stil busy.

The new implementation stays with this behavior as default.
It also allows to flag a request with FLAG_QUEUE_REQUEST which will
(mutex-)wait for the client to get free.

The wait time for threads is limited by 1 sec by default and can be
changed via the client options.

Caveats:
- there is no guaranteed order of execution. The first waiting thread
  to grab the client lock will run.
- there is a slight chance of starvation if the timeout is very long
  but the thread cannot grab the lock.